requirements.txt
requests
pandas
matplotlib
python-dotenv
pyarrow
myfitnesspal
browser_cookie3
